fix: prevent server-only panels from being bundled on client
Split frontpage-panels into two files:
- frontpage-panels.ts (server-only) - contains PANEL_REGISTRY
- frontpage-panels-metadata.ts - contains PANEL_METADATA and defaults

Added 'server-only' directive to prevent panel components from being
bundled on client, which was causing build errors due to 'use cache'
directives in server-only code.

This allows the server component to render panels and pass them to
the client drag-and-drop component without pulling server code into
the client bundle.
